Remove Object type from applyValue functions
This is a test to remove the Object type from the applyValue() functions in Filter and Condition class by replacing the existent method for two new ones to handle String and int return types.
we might want to try and use generics here in the near future.

+++ mbi-data/src/main/java/com/msoft/mbi/data/api/data/filters/Filters.java
@@ -43,7 +43,7 @@ public class Filters {
 
     public int applyValuesInDimension(PreparedStatement stmt, int position) throws BIException {
         if (this.dimensionFilter != null) {
-            position = (Integer) this.dimensionFilter.applyValues(stmt, position);
+            position = this.dimensionFilter.applyValues(stmt, position);
         }
         return position;
     }
